incus-agent: Move to cmd/incus-agent
