+++ cmd/incus-agent/api_1.0.go
@@ -11,11 +11,11 @@ import (
 	"github.com/mdlayher/vsock"
 
 	"github.com/cyphar/incus/client"
-	agentAPI "github.com/cyphar/incus/incus-agent/api"
 	"github.com/cyphar/incus/incus/response"
 	lxdvsock "github.com/cyphar/incus/incus/vsock"
 	"github.com/cyphar/incus/shared"
 	"github.com/cyphar/incus/shared/api"
+	agentAPI "github.com/cyphar/incus/shared/api/agent"
 	"github.com/cyphar/incus/shared/logger"
 	"github.com/cyphar/incus/shared/version"
 )
