+++ cmd/incus-agent/templates.go
